More repo reorganization.
- put dummy code in tests/src/.
- only keep test suite code in tests/src/test.
- Simplify logic in SemanticTests by adding runDiffTest method in SemanticRewriteSuite.
- Move rewrites/*.scala to scalafix-tests/src/resources.
- Remove unused intellij worksheet.
