+++ scalafix-tests/src/main/scala/cats/data/Xor.scala
