+++ scalafix-tests/src/main/scala/cats/implicits.scala
