+++ scalafix-tests/src/main/scala/scala/scalajs/js.scala
