+++ scalafix-tests/src/main/scala/scala/scalajs/js/annotation/JSGlobal.scala
