+++ scalafix-tests/src/main/scala/scala/scalajs/js/annotation/falsealarm/js.scala
