+++ scalafix-tests/src/main/scala/scalafix/test/Generic.scala
