+++ scalafix-tests/src/resources/rewrites/MyRewrite.scala
