+++ scalafix-tests/src/resources/rewrites/MyRewrite2.scala
