+++ scalafix-tests/src/test/resources/checkSyntax/FileRewrite.source
@@ -1,5 +1,5 @@
 rewrites = [
-  "file:rewrites/MyRewrite.scala"
+  "file:scalafix-tests/src/resources/rewrites/MyRewrite.scala"
 ]
 <<< NOWRAP basic
 object Name
