+++ scalafix-tests/src/test/scala/scalafix/tests/ErrorSuite.scala
@@ -1,12 +1,12 @@
-package scalafix.rewrite
+package scalafix.tests
 
 import scalafix.Failure
 import scalafix.Fixed
 import scalafix.Scalafix
-import scalafix.testkit.RewriteSuite
+import scalafix.rewrite.ProcedureSyntax
+import scalafix.testkit.SyntacticRewriteSuite
 
-class ErrorSuite extends RewriteSuite(ProcedureSyntax) {
-
+class ErrorSuite extends SyntacticRewriteSuite(ProcedureSyntax) {
   test("on parse error") {
     val Fixed.Failed(err: Failure.ParseError) = Scalafix.fix("object A {")
   }
